Add option for extra row at end of slickgrid.
This is a modification to the original slickgrid code, so we can't just go replacing slickgrid if slickgrid is updated in the future without making this change. The two lines are commented. Essentially, it just lets us add an unused empty row at the end of the grid to make it work better with some combinations of hardware and software that force a scrollbar at the bottom of the screen which obscures a row because the browser isn't handling it properly.